:recycle: Refactored UnicodeFilterConfiguration
