+++ refarch-backend/src/test/java/de/muenchen/refarch/configuration/filter/nfcconverter/NfcConverterTest.java
@@ -66,14 +66,12 @@ class NfcConverterTest {
     @Mock
     private FilterChain chain;
 
-    private final NfcRequestFilter filter = new NfcRequestFilter();
+    private final NfcRequestFilterConfiguration.NfcRequestFilter filter = new NfcRequestFilterConfiguration.NfcRequestFilter();
 
     // Test that request with configured ContentType is normalized to NFC.
     @Test
     void testFilterIfContenttypeInWhitelist() throws ServletException, IOException {
         mockRequest("text/plain");
-
-        filter.setContentTypes("text/plain;text/html;application/json");
 
         filter.doFilter(req, resp, chain);
 
@@ -93,9 +91,7 @@ class NfcConverterTest {
     // Test that Request not configured ContentType remains unchanged, i.e. is not normalized to NFC.
     @Test
     void testSkipFilterIfContenttypeNotInWhitelist() throws ServletException, IOException {
-        mockRequest("application/postscript");
-
-        filter.setContentTypes("text/plain;text/html");
+        mockRequest("application/notvalid");
 
         filter.doFilter(req, resp, chain);
